+++ matching-service/model/matching-orm.js
@@ -4,9 +4,7 @@ import {
   deleteMatchRequest,
   checkMatchRequestExists,
   findMatchRequest,
-  createMatchSession,
   checkMatchRequestIsMatched,
-  findMatchSession,
   cancelMatchRequest,
 } from './respository.js';
 // need to separate orm functions from repository to decouple business logic from persistence
@@ -117,43 +115,6 @@ export async function ormCheckMatchRequestIsMatched(username, difficulty) {
     return matchFound;
   } catch (err) {
     console.log('ERROR: Error occured when finding users');
-    return { err };
-  }
-}
-
-export async function ormCreateMatchSession(
-  difficulty,
-  username1,
-  username1socketID,
-  username2,
-  username2socketID
-) {
-  try {
-    const newMatchSession = await createMatchSession({
-      difficulty: difficulty,
-      username1: username1,
-      username1socketID: username1socketID,
-      username2: username2,
-      username2socketID: username2socketID,
-    });
-    newMatchSession.save();
-    return newMatchSession;
-  } catch (err) {
-    console.log('ERROR: Could not create new match session');
-    return { err };
-  }
-}
-
-export async function ormFindMatchSession(username, difficulty, username1socketID) {
-  console.log('Running ormFindMatchSession');
-  try {
-    return await findMatchSession({
-      username1: username,
-      difficulty: difficulty,
-      username1socketID: username1socketID,
-    });
-  } catch (err) {
-    console.log('ERROR: Error occured when finding users:', err.body);
     return { err };
   }
 }
